Separate build and sign tx actions + some tweaks
